+++ p21_CPP_Module_09/ex01/RPN.hpp
@@ -11,10 +11,10 @@ class RPN {
   RPN& operator=(const RPN& other);
   ~RPN();
 
-  bool evaluate(const std::string& expr, long long& result) const;
+  bool evaluate(const std::string& expr, int& result) const;
 
  private:
-  bool applyOperator(char op, std::stack<long long>& values) const;
+  bool applyOperator(char op, std::stack<int>& values) const;
 };
 
 #endif
